requirements.txt
FastAPI
Uvicorn
Python-dotenv
Requests